llm-agent-chatbot/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
openai==1.9.0
python-dotenv==1.0.0
duckduckgo-search==8.1.1
pytz==2024.1
requests==2.31.0

